feat(multithread): use (very slow) thread pool
I modified the algorithm to use a more reasonable thread-pooling
approach, but it's still obscenely slow compared to both the
single-threaded variant and the sample selection-sort algorithm (WTF??).

I've commented out the line which puts the algorithm into multithreaded
more for the time-being. Hopefully we can make this a more viable
solution in the next few days.

Signed-off-by: Buckley Ross <brr322@lehigh.edu>
Change-Id: Ic566eea8bbd9312761f3a6517c396d387cbdc6de